+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed Console input in CLI when running under Linux
 - Fixed issue where parallel checks could fail due to UI cross thread access
+- Fixed bug in backup archive trigger (for DLE) when loading tables with dodgy column names (e.g. `[My Group by lolz]`)
 - 
 ...